news_events_dq_assessment/requirements.txt
# Core data handling
pandas>=2.0.0
numpy>=1.24.0

# File formats & config
pyyaml>=6.0
python-dotenv>=1.0.0

# Database support
sqlalchemy>=2.0.0
sqlite-utils>=3.35
psycopg2-binary>=2.9.9   # For PostgreSQL
mysql-connector-python>=8.1.0   # For MySQL

# Dashboard (interactive reports)
dash>=2.11.1
dash-bootstrap-components>=1.6.0
plotly>=5.15.0

# Monitoring & alerts
schedule>=1.2.0
requests>=2.31.0

# Logging
loguru>=0.7.0

# Optional utilities
tqdm>=4.66.1

jsonlines>=3.1.0

pyarrow

fastparquet